requirements.txt
Flask==3.0.3
Flask_SocketIO==5.3.6
pyserial==3.5
python-socketio==5.11.2
